fix: move a ton of functions to separate files
