+++ util/getThemeColor.ts
@@ -1,0 +1,3 @@
+export const getThemeColor = (name: string, theme: any) => {
+  return name.split('.').reduce((o, i) => o[i], theme.colors)
+}
