+++ util/numberWithinRange.ts
@@ -1,0 +1,3 @@
+export const numberWithinRange = (num: number, min: number, max: number) => {
+  return Math.min(Math.max(num, min), max)
+}
